Move Profile service to `service`
